requirements.txt
openai>=1.40.0
python-dotenv>=1.0.1
fastapi>=0.110.0
uvicorn>=0.30.0
pydantic>=2.7.0
python-multipart>=0.0.9
requests>=2.32.2
Pillow>=10.4.0